moved ocr route tests to deparate file
